feat(control-panel): team management on control panel
