+++ api/src/core/database/schemas/events.ts
@@ -6,4 +6,5 @@ export const events = pgTable('events', {
   description: text('description').notNull(),
   createdAt: timestamp('created_at').notNull().defaultNow(),
   updatedAt: timestamp('updated_at').notNull().defaultNow(),
+  deletedAt: timestamp('deleted_at'),
 })
