+++ api/src/features/control-panel/core/ControlPanel.ts
@@ -1,11 +1,21 @@
 import { AppError } from '../../../shared/AppError.ts'
+import type { EventRepository } from '../../event/domain/EventRepository.ts'
+import type { TeamInstanceRepository } from '../../team/domain/TeamInstanceRepository.ts'
 import type { TeamTemplateRepository } from '../../team/domain/TeamTemplateRepository.ts'
 
 export class ControlPanel {
   #teamTemplateRepository: TeamTemplateRepository
+  #eventRepository: EventRepository
+  #teamInstanceRepository: TeamInstanceRepository
 
-  constructor(teamTemplateRepo: TeamTemplateRepository) {
+  constructor(
+    teamTemplateRepo: TeamTemplateRepository,
+    eventRepo: EventRepository,
+    teamInstanceRepo: TeamInstanceRepository
+  ) {
     this.#teamTemplateRepository = teamTemplateRepo
+    this.#eventRepository = eventRepo
+    this.#teamInstanceRepository = teamInstanceRepo
   }
 
   async createTeamTemplate(input: { description?: string; key: string; name: string }) {
@@ -53,5 +63,53 @@ export class ControlPanel {
   async deleteTeamTemplate(id: string) {
     const deleted = await this.#teamTemplateRepository.deleteTeamTemplate(id)
     return deleted
+  }
+
+  async listEvents() {
+    const events = await this.#eventRepository.findAllEvents()
+    return events
+  }
+
+  async getEvent(eventId: string) {
+    const event = await this.#eventRepository.findEvent(eventId)
+
+    if (!event) {
+      throw new AppError('Event not found')
+    }
+
+    return event
+  }
+
+  async createEvent(input: { name: string; description: string }) {
+    const createdEvent = await this.#eventRepository.insertEvent(input)
+    await this.#createEventTeams(createdEvent.id)
+
+    return createdEvent
+  }
+
+  async #createEventTeams(eventId: string) {
+    const teamTemplates = await this.#teamTemplateRepository.listTeamTemplates()
+    const templateIds = teamTemplates.map((template) => template.id)
+    const teamInstances = await this.#teamInstanceRepository.bulkInsertTeamInstances(
+      eventId,
+      templateIds
+    )
+
+    return teamInstances
+  }
+
+  async updateEvent(id: string, input: { name?: string; description?: string }) {
+    const eventToUpdate = await this.#eventRepository.findEvent(id)
+    if (!eventToUpdate) {
+      throw new AppError('Event not found, please check the event id.')
+    }
+
+    const updatedEvent = await this.#eventRepository.updateEvent(id, input)
+    return updatedEvent
+  }
+
+  async deleteEvent(id: string) {
+    const deletedEvent = await this.#eventRepository.softDeleteEvent(id)
+    return deletedEvent
   }
 }
