+++ api/src/features/event/domain/EventRepository.ts
@@ -4,6 +4,6 @@ export interface EventRepository {
   insertEvent: (input: EventInput) => Promise<EventModel>
   findEvent: (id: string) => Promise<EventInput | undefined>
   findAllEvents: () => Promise<EventModel[]>
-  deleteEvent: (id: string) => Promise<EventModel>
+  softDeleteEvent: (id: string) => Promise<EventModel>
   updateEvent: (id: string, input: Partial<EventInput>) => Promise<EventModel>
 }
